mise ne places de autorisation au niveau du user

+++ app/Http/Controllers/ListingController.php
@@ -2,16 +2,31 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Middleware\NotSuspended;
 use App\Http\Requests\StoreListingRequest;
 use App\Http\Requests\UpdateListingRequest;
 use App\Models\Listing;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 use Inertia\Inertia;
 
-class ListingController extends Controller 
+class ListingController extends Controller implements HasMiddleware
 {
+    // pour la gestion des autorisations
+    public static function middleware()
+    {
+        return [
+            new Middleware(
+                ['auth', 'verified', NotSuspended::class],
+                except: ['index', 'show']
+            )
+        ];
+    }
     /**
      * Display a listing of the resource.
      */
@@ -39,6 +54,8 @@ class ListingController extends Controller
      */
     public function create()
     {
+        // le probleme ici est que laravel ne connais quel politque on veut utiliser donc il prend tous les utilisateurs comme suspendu
+        Gate::authorize('create', Listing::class);
         return Inertia::render('Listing/Create');
     }
 
@@ -59,6 +76,8 @@ class ListingController extends Controller
         // transformer le tableau en chaine de caractere
         // $newTags=implode(',',$newTags);
 
+        // le probleme ici est que laravel ne connais quel politque on veut utiliser donc il prend tous les utilisateurs comme suspendu
+        Gate::authorize('create', Listing::class);
         $fields = $request->validate([
             'title' => ['required', 'max:255'],
             'desc' => ['required', 'max:5000'],
@@ -89,9 +108,14 @@ class ListingController extends Controller
      */
     public function show(Listing $listing)
     {
+        Gate::authorize('view', $listing);
         return Inertia::render('Listing/Show', [
             'listing' => $listing,
-            'user' => $listing->user->only(['name', 'id']) //pour limiter les infos sur l'utilisateur au nom et a l'id
+            'user' => $listing->user->only(['name', 'id']), //pour limiter les infos sur l'utilisateur au nom et a l'id
+            // pour pouvoir mettre un cpdition sur l'affichage des boutons
+            'canModify'=>Auth::user() ? Auth::user()->can('modify',$listing) :false
+             //can ici est pareil que Gate
+
         ]);
     }
 
@@ -100,6 +124,7 @@ class ListingController extends Controller
      */
     public function edit(Listing $listing)
     {
+        Gate::authorize('modify',$listing);
         return Inertia::render('Listing/Edit', [
             'listing' => $listing
         ]);
@@ -110,6 +135,7 @@ class ListingController extends Controller
      */
     public function update(Request $request, Listing $listing)
     {
+        Gate::authorize('modify',$listing);
         $fields = $request->validate([
             'title' => ['required', 'max:255'],
             'desc' => ['required', 'max:5000'],
@@ -148,6 +174,8 @@ class ListingController extends Controller
      */
     public function destroy(Listing $listing)
     {
+        Gate::authorize('modify',$listing);
+
         if ($listing->image) {
             Storage::disk('public')->delete($listing->image);
         }
